pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-service"
version = "0.1.0"
description = "Self-hosted RAG service (FastAPI) with Weaviate + Neo4j + Redis + Postgres."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
  "pydantic-settings>=2.5",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "redis>=5.0",
  "httpx>=0.27",
  "structlog>=24.4",
  "weaviate-client>=4.10",
  "neo4j>=5.24",
  "pymupdf>=1.23",
  "tiktoken>=0.8",
  "sentence-transformers>=3.2",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
